requirements.txt
aiohttp>=3.8.1
aiogram>=3.0.0
python-dotenv>=0.20.0
sqlalchemy>=1.4.27
asyncpg>=0.24.0